+++ db/actions/index.js
@@ -1,0 +1,20 @@
+/**
+ * Created by abhishekyadav on 28/06/17.
+ */
+const teachers = require('./teachers');
+const batches = require('./batches');
+const assignments = require('./assignments');
+const students = require('./students');
+const submissions = require('./submissions');
+const users = require('./users');
+const validatePassword = require('./validatepassword');
+
+module.exports = {
+    teachers,
+    batches,
+    assignments,
+    students,
+    submissions,
+    users,
+    validatePassword
+}
