+++ db/db.js
@@ -1,0 +1,61 @@
+/**
+ * Created by varun on 5/24/17.
+ */
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+module.exports = {
+  addStudent,
+  addTeacher,
+  getStudents,
+  getTeachers,
+  searchStudent,
+  searchTeacher,
+  searchStudents,
+  searchTeachers,
+  editStudent,
+  editTeacher,
+  deleteStudent,
+  deleteTeacher,
+  editBatch,
+  addAssignment,
+  getAssignments,
+  searchAssignment,
+  findAssignmentsInBatch,
+  editAssignment,
+  deleteAssignment,
+  addBatch,
+  getBatches,
+  searchBatch,
+  searchBatches,
+  deleteBatch,
+  addSubmission,
+  searchSubmissions,
+  searchByBatch,
+  enrollStudentInBatch,
+  addAssignmentToBatch,
+  endBatch,
+  getSubmissions,
+  acceptSubmissionbyId,
+  acceptSubmissionWithoutId,
+  getAllStudentsInBatch,
+  addTeacher,
+  addUser,
+  addLocalUser,
+  validateLocalPassword
+};
