+++ db/index.js
@@ -1,0 +1,10 @@
+/**
+ * Created by abhishekyadav on 28/06/17.
+ */
+const models = require('./models');
+const actions = require('./actions');
+
+module.exports = {
+    models,
+    actions
+}
