+++ routes/students.js
@@ -2,13 +2,13 @@
  * Created by varun on 5/24/17.
  */
 const express = require('express');
-const db = require('../utils/db');
+const db = require('../db');
 
 const router = express.Router();
 
 //done
 router.post('/new', function(req, res) {
-  db.addUser(data => {
+  db.actions.users.addUser(data => {
     db.addStudent(
       req.body.name,
       req.body.roll,
@@ -31,18 +31,18 @@ router.get('/', function(req, res) {
     type = 'roll';
   } else if (name) {
     type = 'name';
-  }
+  }a
 
   if (type == 'all')
-    db.getStudents(data => {
+    db.actions.students.getStudents(data => {
       res.send(data);
     });
   else if (type == 'roll') {
-    db.searchStudents(roll, type, data => {
+    db.actions.students.searchStudents(roll, type, data => {
       res.send(data);
     });
   } else {
-    db.searchStudents(name, type, data => {
+    db.actions.students.searchStudents(name, type, data => {
       res.send(data);
     });
   }
@@ -50,14 +50,14 @@ router.get('/', function(req, res) {
 
 //done
 router.get('/:id', function(req, res) {
-  db.searchStudent(req.params.id, data => {
+  db.actions.students.searchStudent(req.params.id, data => {
     res.send(data);
   });
 });
 
 //done
 router.put('/:id', function(req, res) {
-  db.editStudent(
+  db.actions.students.editStudent(
     req.params.id,
     req.body.name,
     data => {
@@ -70,7 +70,7 @@ router.put('/:id', function(req, res) {
 
 //TODO add hooks to delete user with it
 router.delete('/:id', function(req, res) {
-  db.deleteStudent(req.params.id, req.query.echo, data => {
+  db.actions.students.deleteStudent(req.params.id, req.query.echo, data => {
     res.send(data);
   });
 });
@@ -78,7 +78,7 @@ router.delete('/:id', function(req, res) {
 //done
 router.post('/:id/enroll/:batchId', function(req, res) {
   let echo = req.query.echo;
-  db.enrollStudentInBatch('id', req.params.id, req.params.batchId, data => {
+  db.actions.batches.enrollStudentInBatch('id', req.params.id, req.params.batchId, data => {
     if (echo) {
       res.send(data);
     } else {
